+++ insights-ui/src/scripts/industry-tariff-reports/tariff-types.ts
@@ -249,3 +249,18 @@ export enum ChartEntityType {
   NEGATIVE_IMPACT = 'NEGATIVE_IMPACT',
   SUMMARY = 'SUMMARY',
 }
+
+/**
+ * Types of report sections supported by this script.
+ */
+export enum ReportType {
+  HEADINGS = 'HEADINGS',
+  UNDERSTAND_INDUSTRY = 'UNDERSTAND_INDUSTRY',
+  TARIFF_UPDATES = 'TARIFF_UPDATES',
+  INDUSTRY_AREA_SECTION = 'INDUSTRY_AREA_SECTION',
+  EVALUATE_INDUSTRY_AREA = 'EVALUATE_INDUSTRY_AREA',
+  REPORT_COVER = 'REPORT_COVER',
+  EXECUTIVE_SUMMARY = 'EXECUTIVE_SUMMARY',
+  FINAL_CONCLUSION = 'FINAL_CONCLUSION',
+  ALL = 'ALL',
+}
